cx_clips: properly use double instead of float for clips time
the ROS 2 time is double, by using float we actually lose significant
precision and it becomes practically unusable.

+++ cx_clips/include/cx_clips/CLIPSEnvManagerNode.h
@@ -114,7 +114,7 @@ private:
   void guarded_load(const std::string &env_name, const std::string &filename);
   CLIPS::Value clips_request_feature(const std::string &env_name,
                                      const std::string &feature_name);
-  float clips_now();
+  double clips_now();
   CLIPS::Values clips_now_systime();
 
   // Function for service clients
